rag_project/requirements.txt
langchain
langchain-community
python-dotenv
streamlit
langchain-groq